+++ artiq/examples/master/repository/arguments_demo.py
@@ -44,6 +44,8 @@ class ArgumentsDemo(EnvExperiment):
         self.setattr_argument("number", NumberValue(42e-6,
                                                     unit="us", scale=1e-6,
                                                     ndecimals=4))
+        self.setattr_argument("integer", NumberValue(42,
+                                                     step=1, ndecimals=0))
         self.setattr_argument("string", StringValue("Hello World"))
         self.setattr_argument("scan", Scannable(global_max=400,
                                                 default=NoScan(325),
@@ -65,7 +67,8 @@ class ArgumentsDemo(EnvExperiment):
         print(self.pyon_value)
         print(self.boolean)
         print(self.enum)
-        print(self.number)
+        print(self.number, type(self.number))
+        print(self.integer, type(self.integer))
         print(self.string)
         for i in self.scan:
             print(i)